pyproject.toml
[project]
name = "git-lfs-gcs"
version = "0.1"
authors = [{name="Igor Kozyrenko", email = "igor@ikseek.com"}]
dependencies = [
    "google-cloud-storage >= 3.0"
]
requires-python = ">=3.9"

[project.scripts]
git-lfs-gcs = "git_lfs_gcs:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"